Fixed the double jump,  its always the same heigth
Double jump used to depend on the velocity of its Y. So i decided to
reset the Y velocity of the rigidbody to 0.0, 0.0 before he's going to
double jump so that the double jump will always have the same heigth as
a normal jump.

Added someground tiles.
Tweaked on the throw function.
Replaced some code to the fixedupdate.